Большое количество измненеий: - Уменьшено количество двигателей до 17; - Добавлены операции откртия и закрытия экрана; - Добавлена светодиодная лента для подсветки картриджей: -- Добалена команда управления светодиодной лентой (цвет светодиода); - Исправлена ошибка, ограничения максимальной длинны сообщения со штрихкодом (было 255 символов); - Увеличен максимальный размер пакета; - Изменен алгоритм сканирования картриджей (с отъездом для чтения штрихкода); - Добавлено управление большим плунжером; - Добавлен парсинг штрихкода картриджа с параметрами анализа и печать в текстовый файл; - Модифицирована модель кассеты с картриджами ( теперь только один картридж в кассете); - Добавлены датчики для определения того, вставлена ли кассета с картриджем в соотвествующий слот; - Добавлено клиент-серверное взаимодействие: -- Добавлен сервер базы данных анализов; -- Добавлен клиент базы данных (эмуляция внесения нового анализа врачем); -- Добавлено окно для внесения анализов из БД (из сервера БД); - Добавлен вывод ячеек ротора;

+++ analyzer_control_mega2560/include/packet_manager.hpp
@@ -6,7 +6,7 @@
 class IPacketListener
 {
 public:
-    virtual void listenPacket(uint8_t *packet, uint8_t packetLength) = 0;
+    virtual void listenPacket(uint8_t *packet, uint16_t packetLength) = 0;
 };
 
 class PacketManager
@@ -18,10 +18,10 @@ public:
     void readPacket();
 
     static void writePacketData(uint8_t byte);
-    static void writePacketData(uint8_t const *bytes, uint8_t bytesNumber);
+    static void writePacketData(uint8_t const *bytes, uint16_t bytesNumber);
     static void writePacketFlag();
 
-    void tryPacketBuild(uint8_t bufferPosition);
+    void tryPacketBuild(uint16_t bufferPosition);
     void findByteStuffingPacket();
 };
